move extract_css logic
